+++ modules/apps/site/site-cms-site-initializer/src/main/resources/META-INF/resources/js/main_view/props_transformer/ContentsFDSPropsTransformer.tsx
@@ -5,18 +5,18 @@
 
 import {IInternalRenderer, IView} from '@liferay/frontend-data-set-web';
 import {openModal} from 'frontend-js-components-web';
+import React from 'react';
 
 import formatActionURL from '../../common/utils/formatActionURL';
+import {ISearchAssetObjectEntry} from '../../structure_builder/types/AssetType';
 import AssetTypeInfoPanel from '../info_panel/AssetTypeInfoPanelContent';
-import {EVENTS} from '../info_panel/util/constants';
-import FilePreviewerModalContent from '../modal/FilePreviewerModalContent';
 import createAssetAction from './actions/createAssetAction';
 import createFolderAction from './actions/createFolderAction';
-import multipleFilesUploadAction from './actions/multipleFilesUploadAction';
 import shareAction from './actions/shareAction';
 import AuthorRenderer from './cell_renderers/AuthorRenderer';
 import NameRenderer from './cell_renderers/NameRenderer';
 import SimpleActionLinkRenderer from './cell_renderers/SimpleActionLinkRenderer';
+import SpaceRenderer from './cell_renderers/SpaceRenderer';
 import TypeRenderer from './cell_renderers/TypeRenderer';
 import addOnClickToCreationMenuItems from './utils/addOnClickToCreationMenuItems';
 import transformViewsItemsProps from './utils/transformViewsItemProps';
@@ -24,13 +24,12 @@ import transformViewsItemsProps from './utils/transformViewsItemProps';
 const ACTIONS = {
 	createAsset: createAssetAction,
 	createFolder: createFolderAction,
-	uploadMultipleFiles: multipleFilesUploadAction,
 };
 
 const OBJECT_ENTRY_FOLDER_CLASSNAME =
 	'com.liferay.object.model.ObjectEntryFolder';
 
-export default function FolderFDSPropsTransformer({
+export default function ContentFDSPropsTransformer({
 	additionalProps,
 	creationMenu,
 	itemsActions = [],
@@ -74,13 +73,23 @@ export default function FolderFDSPropsTransformer({
 					type: 'internal',
 				} as IInternalRenderer,
 				{
+					component: SpaceRenderer,
+					name: 'spaceTableCellRenderer',
+					type: 'internal',
+				} as IInternalRenderer,
+				{
 					component: TypeRenderer,
 					name: 'typeTableCellRenderer',
 					type: 'internal',
 				} as IInternalRenderer,
 			],
 		},
-		infoPanelComponent: () => AssetTypeInfoPanel({additionalProps}),
+		infoPanelComponent: (items: {items: ISearchAssetObjectEntry[]}) => (
+			<AssetTypeInfoPanel
+				additionalProps={additionalProps as any}
+				{...items}
+			/>
+		),
 		itemsActions: itemsActions.map((action) => {
 			if (action?.data?.id === 'actionLink') {
 				return {
@@ -92,28 +101,20 @@ export default function FolderFDSPropsTransformer({
 						),
 				};
 			}
-			else if (action?.data?.id === 'download') {
-				return {
-					...action,
-					isVisible: (item: any) =>
-						Boolean(item?.embedded?.file?.link?.href),
-				};
-			}
 			else if (action?.data?.id === 'view-content') {
 				return {
 					...action,
 					isVisible: (item: any) =>
 						Boolean(
 							item?.entryClassName !==
-								OBJECT_ENTRY_FOLDER_CLASSNAME &&
-								!item?.embedded?.file
+								OBJECT_ENTRY_FOLDER_CLASSNAME
 						),
 				};
 			}
 			else if (action?.data?.id === 'view-file') {
 				return {
 					...action,
-					isVisible: (item: any) => Boolean(item?.embedded?.file),
+					isVisible: () => false,
 				};
 			}
 
@@ -128,10 +129,7 @@ export default function FolderFDSPropsTransformer({
 			event: Event;
 			itemData: any;
 		}) => {
-			if (action?.data?.id === 'show-details') {
-				Liferay.fire(EVENTS.ASSET_DATA, {items: [{...itemData}]});
-			}
-			else if (action?.data?.id === 'share') {
+			if (action?.data?.id === 'share') {
 				const {autocompleteURL, collaboratorURLs} = additionalProps;
 
 				shareAction({
@@ -142,7 +140,7 @@ export default function FolderFDSPropsTransformer({
 					title: itemData.embedded?.title,
 				});
 			}
-			else if (action?.data?.id === 'view-content') {
+			else if (action?.data?.id === 'viewContent') {
 				event?.preventDefault();
 
 				openModal({
@@ -151,22 +149,6 @@ export default function FolderFDSPropsTransformer({
 					url: formatActionURL(itemData, action.href),
 				});
 			}
-			else if (action?.data?.id === 'view-file') {
-				openModal({
-					containerProps: {
-						className: '',
-					},
-					contentComponent: () =>
-						FilePreviewerModalContent({
-							file: itemData.embedded.file,
-							headerName: itemData.embedded.title,
-						}),
-					size: 'full-screen',
-				});
-			}
-		},
-		onSelectedItemsChange: (selectedItems: any[]) => {
-			Liferay.fire(EVENTS.ASSET_DATA, {items: selectedItems});
 		},
 		views: transformViewsItemsProps(views),
 	};
